ajout javascript, code pour groupe (non fonctionnel)

+++ Annuaire/src/annuaire/web/PersonController.java
@@ -1,18 +1,29 @@
 package annuaire.web;
 
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.HashSet;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.propertyeditors.CustomCollectionEditor;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.WebDataBinder;
+import org.springframework.web.bind.annotation.InitBinder;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import annuaire.model.Group;
 import annuaire.model.Person;
+import annuaire.services.IDAOGroup;
 import annuaire.services.IDAOPerson;
 
 @Controller()
@@ -24,7 +35,10 @@ public class PersonController {
 
 	@Autowired
 	IDAOPerson daoPerson;
-	
+
+	@Autowired
+	IDAOGroup daoGroup;
+
 	@RequestMapping(value = "/annuaire.htm")
 	public String list() {
 		return "listPerson";
@@ -37,7 +51,7 @@ public class PersonController {
 
 	@RequestMapping(value = "/detail.htm", method = RequestMethod.GET)
 	public String detailPerson(@ModelAttribute Person p) {
-		
+
 		return "person";
 	}
 
@@ -62,6 +76,7 @@ public class PersonController {
 		p.setWebsite("");
 		p.setDateOfBirth(null);
 		p.setPassword("");
+	//	p.setGroups(new HashSet<Group>());
 		return p;
 	}
 
@@ -72,7 +87,7 @@ public class PersonController {
 			return "redirect:annuaire.htm";
 		}
 
-		if(!p.getLogin().equals(user.getLogin())) {
+		if(!p.getLogin().equals(user.getLogin()) && !p.getGroups().contains("ADMINISTRATEUR")) {
 			return "redirect:annuaire.htm";
 		}
 
@@ -84,12 +99,12 @@ public class PersonController {
 		if(result.hasErrors()) {
 			return "personForm";
 		}
-		
+
 		if(p == null || p.getLogin().isEmpty()) {
 			return "redirect:annuaire.htm";
 		}
 
-		if(!p.getLogin().equals(user.getLogin())) {
+		if(!p.getLogin().equals(user.getLogin()) && !p.getGroups().contains("ADMINISTRATEUR")) {
 			return "redirect:annuaire.htm";
 		}
 
@@ -98,8 +113,9 @@ public class PersonController {
 		return "redirect:detail.htm?id=" + p.getLogin();
 	}
 
-	@RequestMapping(value = "/add.htm")
+	@RequestMapping(value = "/add.htm", method = RequestMethod.GET)
 	public String addPerson() {
+
 		return "personAddition";
 	}
 
@@ -107,13 +123,22 @@ public class PersonController {
 	public String saveNewPerson(@ModelAttribute @Valid Person p, BindingResult result) {
 
 		if(result.hasErrors()) {
+			System.out.println(result.getFieldError());
 			return "personAddition";
 		}
 
-		daoPerson.addPerson(p);
 
-		return "redirect:detail.htm?id=" + p.getLogin();
+//		for(Group g : user.getGroups()){
+//			if (g.getGroupname().equals("ADMINISTRATEUR")){
+				daoPerson.addPerson(p);
+				return "redirect:detail.htm?id=" + p.getLogin();
+//			}
+//		}
+
+		//return "redirect:annuaire.htm";
 	}
+
+
 
 	@RequestMapping(value = "/delete.htm", method = RequestMethod.GET)
 	public String deletePerson(@ModelAttribute Person p) {
@@ -122,8 +147,53 @@ public class PersonController {
 			return "redirect:annuaire.htm";
 		}
 
+		if(!p.getGroups().contains("ADMINISTRATEUR")) {
+			return "redirect:annuaire.htm";
+		}
+
 		daoPerson.deletePerson(p.getLogin());
 
 		return "redirect:annuaire.htm";
 	}
+
+	//	@ModelAttribute("groupsList")
+	//	public Map<Set<Group>, String> productTypes() {
+	//		
+	//		Collection<Group> groups = daoGroup.findAllGroups();
+	//		
+	//	    Map<Set<Group>,String> types = new LinkedHashMap<Set<Group>,String>();
+	//	    
+	//	    for(Group g: groups){
+	//	    	Set<Group> tmpSG = new HashSet<Group>();
+	//	    	tmpSG.add(g);
+	//	    	types.put(tmpSG, g.getGroupname());
+	//	    }
+	//	    return types;
+	//	}
+
+	@ModelAttribute("groupsList")
+	public List<Group> productTypes() {
+
+		return (List<Group>) daoGroup.findAllGroups();
+	}
+
+
+
+	@InitBinder
+	protected void initBinder(WebDataBinder binder) {
+		binder.registerCustomEditor(Set.class, "groups", new CustomCollectionEditor(Set.class)
+		{
+			@Override
+			protected Object convertElement(Object element)
+			{
+
+				System.out.println("-------------"+(String) element);
+
+				Set<Group> retour = new HashSet<Group>();
+				retour.add(daoGroup.findGroup((String) element));
+				return retour;
+			}
+		});
+	}
+
 }
